Closing this project for good.
Originally, I wanted to embed the last.fm player in my swing GUI, but
that's not possible. I wanted to lookup songs on services like
grooveshark and then embed their player, but that's way too costly. Then
I wanted to just create the playlist for a user's mood in their last.fm
account, but last.fm authentication API does not seem like a good one.

They do not callback to let us know if the user has completed the
authentication. Its broken! One option is to add a button in the GUI
that user can use to let me know that he has completed the
authentication, which is bad design, and I refuse to implement that (for
now).

Hence, I'm dropping this project for good (for now).